ecp_ZZZ: Use correct acceptance criterion in `fromhash`.
Chen and Li's paper (whose appendix we're using as the definition of our
G1-point-from-a-hash function), as well as the FIDO v2 spec, specify to
set the y-coordinate of the returned point to the *minimum* of the
y-coordinate of the generated point and the modulus-minus-y-coord.
However, until now, we have been choosing between these two curve points
based on their parity (because I was using the "Hunting and Pecking in
ECC groups" method from the Dragonfly spec; I don't remember now why I
was using that...).